Update remote-method.xml
There is a lot of confusion surrounding proper usage of the `remote` method based on the number of Stack Overflow questions about it.  I am proposing this slight change to help clarify.

Proposed:

The serverside response must be a JSON string that must be `"true"` for valid elements, and can be `"false"`, `undefined`, or `null` for invalid elements, using the default error message.  If the serverside response is a string, eg. `"That name is already taken, try peter123 instead"`, this string will be displayed as a custom error message in place of the default.