Entries: Changed generic form selector
Some examples on this page are using `$(".selector").validate()` and others are using `$("#myform").validate()`.

Using `$(".selector")` implies that `.validate()` can be directly attached to a class selector where, if this is done without using an `.each()`, it would only be attached to the first instance of this class.

The `$(".selector")` should be changed into `$("#myform")` on the entire page for better consistency and less confusion.  I think `$("#myform")` is more appropriate since the `.validate()` method should only be attached to a selector that represents a single `form` element.

Closes #24.