Update reference.md
Clarified this section to reflect many questions on StackOverflow.   It's a common mistake to leave out or duplicate the `name` attribute.  It's also a common misconception that an `id` is needed at all for this plugin to work.  Nowhere in the documentation is it explicitly stated that `name` attribute must be unique... only implied.  My proposed changes spells these items out more clearly.

The most important markup requirements are moved to the very top of this section in a paragraph preceded by the word "Mandated".  Removed the triple single quotes from the word "required" as this just seems like we're talking about the `required` rule.

And to eliminate any confusion, anything that has no effect on the operation of the plugin is confined to a paragraph preceded by the word "Optionally".