Reference: clarified markup recommendations.
Clarified this section to reflect many questions on StackOverflow.
It's a common mistake to leave out or duplicate the `name` attribute.
It's also a common misconception that an `id` is needed at all for this
plugin to work. Nowhere in the documentation is it explicitly stated that
`name` attribute must be unique...only implied.

Closes #25.