Entries: document onclick for select elements
When answering a question on StackOverflow about a broken `select` element, the OP had this option set to `false`.  The `select` element was only being validated on focusout and on submit.  Selecting items had no effect on validation message of this element.   Removing this option, since it cannot be set to `true`, fixed the issue.  Therefore, the documentation is incomplete.  This option affects validation of `select` elements as well as radio and checkbox.

https://jsfiddle.net/pu752c3a/